requirements.txt
spotipy==2.23.0
pandas>=2.0
rapidfuzz>=3.6
python-dotenv>=1.0
PyYAML>=6.0
